requirements.txt
requests==2.22.0
python-dotenv==0.12.0
python-telegram-bot==11.1.0
dialogflow==1.0.0
vk-api==11.8.0